pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "2.7.2"
apache-airflow-providers-amazon = "8.7.1"
pendulum = "2.1.2"
connexion = "2.14.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
